ENH: created separate F10.7 methods
Created separate F10.7 methods, updating acknowledgements and references, and removed old combined methods file.
